# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Stakeframe API",
    "version": "0.0.0",
    "description": "Aplicação pessoal com login Google exclusivo do proprietário. Operações financeiras exigem sessão, origem, chave de idempotência e versão. Decimais são strings; erros usam código estável e requestId gerado pelo servidor."
  },
  "components": {
    "securitySchemes": {
      "localOwnerSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "stakeframe.session_token",
        "description": "Cookie HttpOnly de sessão, usado somente no HTTP de loopback. Não inserir tokens Google aqui."
      },
      "secureOwnerSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Secure-stakeframe.session_token",
        "description": "Variante Secure para HTTPS; infraestrutura HTTPS ainda não publicada."
      }
    },
    "schemas": {
...
      "ImportPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "source": { "type": "string", "enum": ["web", "telegram"] },
                "caption": { "type": "string" },
                "state": {
                  "type": "string",
                  "enum": ["pending", "processing", "review", "failed", "discarded", "imported"]
                },
                "version": {
                  "type": "integer",
                  "minimum": 0,
                  "exclusiveMinimum": true,
                  "maximum": 9007199254740991
                },
                "attempts": { "type": "integer", "minimum": 0, "maximum": 9007199254740991 },
                "createdAt": {
                  "type": "string",
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                },
                "updatedAt": {
                  "type": "string",
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                },
                "errorCode": { "nullable": true, "type": "string" },
                "betId": {
                  "nullable": true,
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "imageAvailable": { "type": "boolean" }
              },
              "required": [
                "id",
                "source",
                "caption",
                "state",
                "version",
                "attempts",
                "createdAt",
                "updatedAt",
                "errorCode",
                "betId",
                "imageAvailable"
              ],
              "additionalProperties": false
            }
          },
          "total": { "type": "integer", "minimum": -9007199254740991, "maximum": 9007199254740991 },
          "page": { "type": "integer", "minimum": -9007199254740991, "maximum": 9007199254740991 },
          "pageSize": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          }
        },
        "required": ["items", "total", "page", "pageSize"],
        "additionalProperties": false
      },
...
      "ImportDetail": {
        "type": "object",
        "properties": {
          "item": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "source": { "type": "string", "enum": ["web", "telegram"] },
              "caption": { "type": "string" },
              "state": {
                "type": "string",
                "enum": ["pending", "processing", "review", "failed", "discarded", "imported"]
              },
              "version": {
                "type": "integer",
                "minimum": 0,
                "exclusiveMinimum": true,
                "maximum": 9007199254740991
              },
              "attempts": { "type": "integer", "minimum": 0, "maximum": 9007199254740991 },
              "createdAt": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
              },
              "updatedAt": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
              },
              "errorCode": { "nullable": true, "type": "string" },
              "betId": {
                "nullable": true,
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "imageAvailable": { "type": "boolean" }
            },
            "required": [
              "id",
              "source",
              "caption",
              "state",
              "version",
              "attempts",
              "createdAt",
              "updatedAt",
              "errorCode",
              "betId",
              "imageAvailable"
            ],
            "additionalProperties": false
          },
          "extraction": {
            "nullable": true,
            "type": "object",
            "properties": {
              "bookmaker": { "nullable": true, "type": "string", "minLength": 1, "maxLength": 500 },
              "reference": { "nullable": true, "type": "string", "minLength": 1, "maxLength": 500 },
              "placedAtText": {
                "nullable": true,
                "type": "string",
                "minLength": 1,
                "maxLength": 500
              },
              "currency": { "nullable": true, "type": "string", "enum": ["BRL"] },
              "stake": {
                "nullable": true,
                "type": "string",
                "pattern": "^(0|[1-9]\\d{0,11})(\\.\\d{1,4})?$"
              },
              "odds": {
                "nullable": true,
                "type": "string",
                "pattern": "^(0|[1-9]\\d{0,11})(\\.\\d{1,4})?$"
              },
              "potentialReturn": {
                "nullable": true,
                "type": "string",
                "pattern": "^(0|[1-9]\\d{0,11})(\\.\\d{1,4})?$"
              },
              "freebet": { "nullable": true, "type": "boolean" },
              "selections": {
                "minItems": 1,
                "maxItems": 40,
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "event": {
                      "nullable": true,
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "sport": {
                      "nullable": true,
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "market": {
                      "nullable": true,
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "selection": {
                      "nullable": true,
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "odds": {
                      "nullable": true,
                      "type": "string",
                      "pattern": "^(0|[1-9]\\d{0,11})(\\.\\d{1,4})?$"
                    },
                    "eventDateText": {
                      "nullable": true,
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    }
                  },
                  "required": ["event", "sport", "market", "selection", "odds", "eventDateText"],
                  "additionalProperties": false
                }
              },
              "warnings": {
                "maxItems": 20,
                "type": "array",
                "items": { "type": "string", "minLength": 1, "maxLength": 500 }
              }
            },
            "required": [
              "bookmaker",
              "reference",
              "placedAtText",
              "currency",
              "stake",
              "odds",
              "potentialReturn",
              "freebet",
              "selections",
              "warnings"
            ],
            "additionalProperties": false
          },
          "labels": {
            "type": "object",
            "properties": {
              "tipster": { "nullable": true, "type": "string" },
              "bookmaker": { "nullable": true, "type": "string" },
              "requiresReview": { "type": "boolean" }
            },
            "required": ["tipster", "bookmaker", "requiresReview"],
            "additionalProperties": false
          },
          "matches": {
            "type": "object",
            "properties": {
              "tipsterId": {
                "nullable": true,
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "captionBookmakerId": {
                "nullable": true,
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "extractedBookmakerId": {
                "nullable": true,
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "conflict": { "type": "boolean" }
            },
            "required": ["tipsterId", "captionBookmakerId", "extractedBookmakerId", "conflict"],
            "additionalProperties": false
          },
          "duplicates": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "betId": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "reference": { "type": "string" },
                "bookmakerId": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "stake": { "type": "string" },
                "placedAt": {
                  "type": "string",
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                },
                "reasons": {
                  "type": "array",
                  "items": { "type": "string", "enum": ["image", "reference", "similar"] }
                }
              },
              "required": ["betId", "reference", "bookmakerId", "stake", "placedAt", "reasons"],
              "additionalProperties": false
            }
          },
          "duplicateCount": { "type": "integer", "minimum": 0, "maximum": 9007199254740991 },
          "automatic": { "type": "boolean", "enum": [false] },
          "automaticReason": { "type": "string", "enum": ["LAYOUT_NOT_VALIDATED"] }
        },
        "required": [
          "item",
          "extraction",
          "labels",
          "matches",
          "duplicates",
          "duplicateCount",
          "automatic",
          "automaticReason"
        ],
        "additionalProperties": false
      },
...
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "NOT_FOUND",
                  "INVALID_REQUEST",
                  "INTERNAL_ERROR",
                  "AUTH_NOT_CONFIGURED",
                  "UNAUTHENTICATED",
                  "ORIGIN_NOT_ALLOWED",
                  "AUTH_REQUEST_FAILED",
                  "RATE_LIMITED",
                  "AUTH_UNAVAILABLE",
                  "STATE_CONFLICT",
                  "VERSION_CONFLICT",
                  "IDEMPOTENCY_CONFLICT",
                  "INVALID_FINANCIAL_OPERATION",
                  "UNIT_REQUIRED",
                  "NOT_INITIALIZED",
                  "ALIAS_CONFLICT",
                  "DUPLICATE_REVIEW_REQUIRED",
                  "INVALID_INBOX_IMAGE",
                  "INBOX_BUSY",
                  "INBOX_CAPACITY_REACHED",
                  "ATTACHMENT_UNAVAILABLE"
                ]
              },
              "message": { "type": "string" },
              "requestId": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              }
            },
            "required": ["code", "message", "requestId"],
            "additionalProperties": false
          }
        },
        "required": ["error"],
        "additionalProperties": false
      },
...
    }
  },
  "paths": {
...
    "/api/v1/imports": {
      "get": {
        "operationId": "listImports",
        "summary": "Consultar importações privadas",
        "tags": ["Importações"],
        "parameters": [
          {
            "schema": { "default": 1, "type": "integer", "minimum": 1, "maximum": 100000 },
            "in": "query",
            "name": "page",
            "required": false
          },
          {
            "schema": { "default": 25, "type": "integer", "minimum": 1, "maximum": 100 },
            "in": "query",
            "name": "pageSize",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": ["pending", "processing", "review", "failed", "discarded", "imported"]
            },
            "in": "query",
            "name": "state",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "query",
            "name": "betId",
            "required": false
          }
        ],
        "security": [{ "localOwnerSession": [] }, { "secureOwnerSession": [] }],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ImportPage" } }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "413": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "default": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          }
        }
      },
      "post": {
        "operationId": "uploadTicket",
        "summary": "Enviar comprovante para revisão",
        "tags": ["Importações"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": {
                    "type": "string",
                    "minLength": 4,
                    "maxLength": 11184812,
                    "pattern": "^[A-Za-z0-9+/]+={0,2}$"
                  },
                  "caption": { "type": "string", "maxLength": 1024 }
                },
                "required": ["image", "caption"],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "header",
            "name": "idempotency-key",
            "required": true
          }
        ],
        "security": [{ "localOwnerSession": [] }, { "secureOwnerSession": [] }],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    }
                  },
                  "required": ["id"],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "413": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "default": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          }
        }
      }
    },
    "/api/v1/imports/{id}": {
      "get": {
        "operationId": "getImport",
        "summary": "Conferir extração, aliases e possíveis duplicações",
        "tags": ["Importações"],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [{ "localOwnerSession": [] }, { "secureOwnerSession": [] }],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ImportDetail" } }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "413": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "default": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          }
        }
      }
    },
    "/api/v1/imports/{id}/image": {
      "get": {
        "operationId": "getImportImage",
        "summary": "Ler comprovante privado com sessão do proprietário",
        "tags": ["Importações"],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [{ "localOwnerSession": [] }, { "secureOwnerSession": [] }],
        "responses": {
          "200": {
            "description": "Bytes privados; sessão exigida em cada leitura.",
            "content": {
              "image/png": { "schema": { "type": "string", "format": "binary" } },
              "image/jpeg": { "schema": { "type": "string", "format": "binary" } }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "413": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "default": {
            "description": "Default Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          }
        }
      }
    }
  },
  "servers": [{ "url": "/", "description": "Mesma origem da aplicação" }],
...
}